# src/main/resources/swagger.yml
openapi: 3.0.0
info:
  title: Speak Up Complain APP
  version: 1.0.0
servers:
  - url: http://127.0.0.1:8080
components:
  securitySchemes:
    noauthAuth:
      type: http
      scheme: noauth
    bearerAuth:
      type: http
      scheme: bearer
...
paths:
...
  /api/v1/complains:
    get:
      tags:
        - Complains
      summary: users complain list
      security:
        - bearerAuth: []
      parameters:
        - name: Accept
          in: header
          schema:
            type: string
          example: application/json
      responses:
        '200':
          description: Successful response
          content:
            application/json: {}
    post:
      tags:
        - Complains
      summary: create complain
      requestBody:
        content:
          application/json:
            schema:
              type: object
              example:
                title: Food Complain
                description: No Description
      security:
        - bearerAuth: []
      parameters:
        - name: Accept
          in: header
          schema:
            type: string
          example: application/json
      responses:
        '200':
          description: Successful response
          content:
            application/json: {}
...